requirements.txt
django==1.10.2
djangorestframework==3.4.7
django-webpack-loader==0.3.3
psycopg2==2.6.2
fabric3==1.12.post1
